vkprofiles: Refactor to move reusable code in utils files
